+++ deps/getline/bufreader.h
@@ -20,7 +20,7 @@ class bufreader {
 	bool done = false;
 	int linesize = 0;
 	char* getline();
-	int addline();
+	long long addline();
 	bool addrefresh(int);
 	bufreader(){}
 	~bufreader(){if (f) fclose(f);}
